+++ providers/__init__.py
@@ -1,6 +1,11 @@
 from .base import DeviceStatus, Provider  # noqa: F401
+from .audeze import AudezeProvider  # noqa: F401
 from .razer import RazerProvider  # noqa: F401
 from .wlmouse import WLmouseProvider  # noqa: F401
+from .mchose import MchoseProvider  # noqa: F401
 from .bluetooth import BluetoothProvider  # noqa: F401
 from .xinput import XInputProvider  # noqa: F401
+from .logitech import LogitechProvider  # noqa: F401
+from .steelseries import SteelSeriesProvider  # noqa: F401
+from .hyperx import HyperXProvider  # noqa: F401
 from .playstation import PlayStationProvider  # noqa: F401